+++ demo/pages/notfound.nim
@@ -5,8 +5,5 @@ import kraut/context
 
 proc render*(context: Context): VNode =
   buildHtml(tdiv):
-    p:
-      text "This is index"
-    ul:
-      li: a(href = "#/users/"): text "Users"
+    h1: text "Not found"