+++ demo/pages/users.nim
@@ -1,9 +1,14 @@
-import karax/[karaxdsl, vdom]
+import std/strformat
+import karax/[karaxdsl, kbase, vdom]
 
 import kraut/context
 
 
 proc render*(context: Context): VNode =
   buildHtml(tdiv):
-    text "This is users page"
+    p:
+      text "This is users page"
+    ul:
+      for userId in ["123", "456", "789"]:
+        li: a(href = kstring fmt"#/users/{userId}/"): text fmt"User {userId}"