+++ changelog.md
@@ -8,6 +8,12 @@
 -   [d]—docs improvement
 
 
+## 1.0.1 (March 7, 2023)
+
+-   [+] Add `defaultRenderer` param to `routeRenderer` to handle the Not Found situation.
+-   [+] Add `kraut/sugar` module with `routes` macro to generate routes table from module names.
+
+
 ## 1.0.0 (February 27, 2023)
 
 -   🎉 initial release.